- remove preview Play app - rename package
