+++ facade/src/main/scala/org/scalajs/threejs/extensions/controls/CameraControls.scala
@@ -1,0 +1,20 @@
+package org.scalajs.threejs.extensions.controls
+
+import org.scalajs.dom.MouseEvent
+
+trait CameraControls {
+
+  def onMouseDown(event: MouseEvent): Unit
+
+  def onMouseMove(event: MouseEvent): Unit
+
+  def onMouseUp(event: MouseEvent): Unit
+
+  def onMouseWheel(event: MouseEvent): Unit
+
+  def update(): Unit
+
+  def enabled: Boolean
+
+}
+
