+++ index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="UTF-8">
+    <title>The ScalaJS ThreeJS Facade Preview</title>
+    <style>
+        html, body {
+            margin: 0;
+            overflow: hidden;
+        }
+    </style>
+    <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/three.js/77/three.js"></script>
+</head>
+<body>
+<!-- Include Scala.js compiled code -->
+<script type="text/javascript" src="./preview/target/scala-2.12/preview-fastopt.js"></script>
+</body>
+</html>
