--- preview/backend/src/main/scala/org/denigma/preview/templates/MyStyles.scala
@@ -1,13 +1,0 @@
-package org.denigma.preview.templates
-
-import scalacss.Defaults._
-
-object MyStyles extends StyleSheet.Standalone {
-  import dsl._
-
- /* "#container" - (
-    width(1024 px),
-    height(768 px)
-    )
-*/
-}
