--- preview/backend/src/main/twirl/examples.scala.html
@@ -1,16 +1,0 @@
-<section class="ui blue piled message">
-    ScalaJS interface for THREE.js javascript lib. You can rotate the cube and sprite by holding left mouse button
-</section>
-<section id="container">
-
-</section>
-<section class="ui blue segment">
-    <div class="ui orange segment">
-        <div class="ui medium header">ScalaJS code</div>
-        <textarea id="scala"> </textarea>
-    </div>
-    <div class="ui teal segment">
-        <div class="ui medium header">Html code</div>
-        <textarea id="html"> </textarea>
-    </div>
-</section>>
