--- preview/backend/src/main/twirl/libs.scala.html
@@ -1,17 +1,0 @@
-<link rel="stylesheet" href="/styles/mystyles.css" />
-
-<script type="text/javascript" src="/lib/jquery/jquery.min.js"></script>
-
-<link rel="stylesheet" href="/lib/Semantic-UI/semantic.css" />
-<script type="text/javascript" src="/lib/Semantic-UI/semantic.js"></script>
-
-<link rel="stylesheet" href="/lib/codemirror/lib/codemirror.css" type="text/css">
-<script type="text/javascript" src="/lib/codemirror/lib/codemirror.js"></script>
-<script type="text/javascript" src="/lib/codemirror/mode/xml/xml.js"></script>
-<script type="text/javascript" src="/lib/codemirror/mode/htmlmixed/htmlmixed.js"></script>
-<script type="text/javascript" src="/lib/codemirror/mode/clike/clike.js"></script>
-
-<script type="text/javascript" src="/lib/three.js/three.js"></script>
-
-<script src="/resources/OrbitControls.js"></script>
-<script src="/resources/CSS3DRenderer.js"></script>
