--- preview/backend/src/main/twirl/main.scala.html
@@ -1,4 +1,0 @@
-@(content:Option[Html] = None)
-<main id="main" class="ui green tertiary inverted segment">
-  @{content.getOrElse(examples())}
-</main>
