--- preview/backend/src/main/twirl/runner.scala.html
@@ -1,10 +1,0 @@
-@(into:String = "main",leftSidebar:Boolean = true)
-<script src="/resources/frontend-fastopt.js"></script>
-<script>
-$( document ).ready(function() {
-    FrontEnd().main(); //activate ScalaJS main class
-    @if(leftSidebar) {
-    FrontEnd().showLeftSidebar(); //activate sidebar (if exists)
-    }
-});
-</script>
