--- preview/backend/src/main/twirl/sidebar.scala.html
@@ -1,15 +1,0 @@
-<section class="ui left  labeled sidebar green segment" data-view="sidebar">
-
-  <div class="ui breadcrumb">
-    <i class="right arrow icon divider"></i>
-    <header class="ui section header" data-bind="title"></header>
-    <i class="right arrow icon divider"></i>
-  </div>
-  <div class="item">
-    <img  class="ui image" data-bind-src="logo">
-  </div>
-
-</section>
-
-
-
