--- preview/backend/src/universal/conf/application.ini
@@ -1,4 +1,0 @@
--J-Xms256M
--J-Xmx8192M
-
--jvm-debug 9999
